Move aggregated metadata from SWORD to Protocol, to have it available for Zenodo
