+++ deposit/tests/test_protocol.py
@@ -138,6 +138,29 @@ class MetaTestProtocol():
         assert isinstance(form, Form)
 
 
+    def test_get_prefered_records(self, dummy_paper):
+        """
+        Tests that only records from CrossRef and BASE are in the list
+        """
+        self.protocol.paper = dummy_paper
+        crossref = OaiSource.objects.get(identifier='crossref')
+        base = OaiSource.objects.get(identifier='base')
+        zenodo = OaiSource.objects.get(identifier='zenodo')
+        for i in range(3):
+            for source in [crossref, base, zenodo]:
+                OaiRecord.objects.create(
+                    about=dummy_paper,
+                    identifier=source.identifier + str(i),
+                    source=source
+                )
+        prefered_records = self.protocol._get_prefered_records()
+        assert len(prefered_records) == 6
+        for i in range(3):
+            assert prefered_records[i].source == crossref
+        for i in range(3,6):
+            assert prefered_records[i].source == base
+
+
     def test_init_deposit(self, user_isaac_newton, book_god_of_the_labyrinth):
         """
         init_deposit shall return a bool
